move stores directory
